Update GetUnmatched.py
Found the problem with the list iteration of the field names. The script was iterating the list while it was modifying it, so it would remove an element, which would take it to the next element, and then it would hit the end of the loop, causing it to move another element over. This would skip any element that occurred after an element that should be removed. Fixed by creating a copy of the list to iterate through.